FESI1-37 [Refactor]: 이동된 파일 themeList를 사용하는 파일의 경로 수정
