+++ src/components/gathering/ThemeSelector.tsx
@@ -1,6 +1,6 @@
 import Input from '@/components/@shared/Input';
 import Button from '@/components/@shared/Button';
-import { themeList } from '@/data/themeList';
+import { themeList } from '@/constants/themeList';
 
 interface ThemeSelectorProps {
   location: string;
